Refactor `test_ops`
